CRM/crm-core/requirements.txt
alabaster==1.0.0
alembic==1.18.3
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
babel==2.18.0
bcrypt==4.0.1
bootstrap-py==2.0.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.4
defusedxml==0.7.1
dnspython==2.8.0
docutils==0.22.4
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.128.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
idna==3.11
imagesize==1.4.1
Jinja2==3.1.6
jose==1.0.0
Mako==1.3.10
MarkupSafe==3.0.3
numpy==2.4.1
packaging==26.0
pandas==3.0.0
passlib==1.7.4
pguard==0.4.0
psycopg2-binary==2.9.11
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
requests==2.32.5
roman-numerals==4.1.0
rsa==4.9.1
setuptools==80.10.2
setuptools-scm==9.2.2
six==1.17.0
smmap==5.0.2
snowballstemmer==3.0.1
Sphinx==9.1.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.46
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
